refactor: 优化用户状态字段类型及查询逻辑
- 将 `UserDO` 和 `UserDTO` 中的 `status` 字段类型从 `Integer` 改为 `Short`
- 在 `TeacherDAO` 中简化了 `getTeacherList` 方法的参数处理逻辑，移除了不必要的变量和空检查
- 更新了 `TeacherMapper` 接口及其 XML 文件中的 SQL 语句，使用 `Integer` 类型的 `status` 参数，并优化了 SQL 语句结构
- 在 `TeacherLogic` 中更新了 `setStatus` 方法的逻辑，确保 `status` 字段正确设置
- 修改了 `TeacherTest` 中的测试方法参数，以适应新的参数类型
- 更新了 `UserEditVO` 中的 `status` 字段类型为 `Short`

+++ src/main/java/com/frontleaves/scheduling/models/entity/UserDO.java
@@ -83,7 +83,7 @@ public class UserDO {
     /**
      * 用户状态 0: 禁用 1: 启用
      */
-    private Integer status;
+    private Short status;
 
     /**
      * 用户是否被封禁 0: 未封禁 1: 已封禁
